# backend/openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SignInDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "refresh-token": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    }
  },
  "info": {
    "title": "Yet another TIMDB application",
    "description": "Api",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refresh_token"
      },
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "access_token",
        "in": "header"
      }
    },
    "schemas": {
...
      "SignInDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}