# backend/src/modules/api/auth/dto/sign-in.dto.ts
import { IsString, Length, MinLength } from 'class-validator';

export class SignInDto {
  @IsString()
  @Length(5, 100)
  username!: string;

  @IsString()
  @MinLength(12)
  password!: string;
}

export class SignInResponseDto {
  accessToken!: string;
  expiresAt!: Date;
}